[IHR2-6577] - refactor
