+++ service/openapi/src/main/java/org/dotwebstack/framework/service/openapi/response/ResponseObject.java
@@ -21,5 +21,4 @@ public class ResponseObject {
         .isEmpty();
   }
 
-
 }
